+++ backend/apid/apid.go
@@ -41,6 +41,7 @@ type APId struct {
 	tls                 *types.TLSOptions
 	cluster             clientv3.Cluster
 	etcdClientTLSConfig *tls.Config
+	clusterVersion      string
 }
 
 // Option is a functional option.
@@ -57,6 +58,7 @@ type Config struct {
 	Cluster             clientv3.Cluster
 	EtcdClientTLSConfig *tls.Config
 	Authenticator       *authentication.Authenticator
+	ClusterVersion      string
 }
 
 // New creates a new APId.
@@ -73,6 +75,7 @@ func New(c Config, opts ...Option) (*APId, error) {
 		cluster:             c.Cluster,
 		etcdClientTLSConfig: c.EtcdClientTLSConfig,
 		Authenticator:       c.Authenticator,
+		clusterVersion:      c.ClusterVersion,
 	}
 
 	// prepare TLS configs (both server and client)
@@ -101,7 +104,7 @@ func New(c Config, opts ...Option) (*APId, error) {
 	router := mux.NewRouter().UseEncodedPath()
 	router.NotFoundHandler = middlewares.SimpleLogger{}.Then(http.HandlerFunc(notFoundHandler))
 	router.Handle("/metrics", promhttp.Handler())
-	registerUnauthenticatedResources(router, a.store, a.cluster, a.etcdClientTLSConfig)
+	registerUnauthenticatedResources(router, a.store, a.cluster, a.etcdClientTLSConfig, a.clusterVersion)
 	a.registerGraphQLService(router, c.URL, tlsClientConfig)
 	registerAuthenticationResources(router, a.store, a.Authenticator)
 	a.registerRestrictedResources(router)
@@ -186,6 +189,7 @@ func registerUnauthenticatedResources(
 	store store.Store,
 	cluster clientv3.Cluster,
 	etcdClientTLSConfig *tls.Config,
+	clusterVersion string,
 ) {
 	mountRouters(
 		NewSubrouter(
@@ -194,6 +198,7 @@ func registerUnauthenticatedResources(
 			middlewares.LimitRequest{},
 		),
 		routers.NewHealthRouter(actions.NewHealthController(store, cluster, etcdClientTLSConfig)),
+		routers.NewVersionRouter(actions.NewVersionController(clusterVersion)),
 	)
 }