+++ backend/etcd/etcd.go
@@ -31,6 +31,14 @@ const (
 	// EtcdStartupTimeout is the amount of time we give the embedded Etcd Server
 	// to start.
 	EtcdStartupTimeout = 60 // seconds
+
+	// DefaultMaxRequestBytes is the default maximum request size for etcd
+	// requests (1.5 MB)
+	DefaultMaxRequestBytes = 1.5 * (1 << 20)
+
+	// DefaultQuotaBackendBytes is the default database size limit for etcd
+	// databases (4 GB)
+	DefaultQuotaBackendBytes = 1 << 32
 )
 
 func init() {
@@ -51,6 +59,11 @@ type Config struct {
 
 	ClientTLSInfo TLSInfo
 	PeerTLSInfo   TLSInfo
+
+	CipherSuites []string
+
+	MaxRequestBytes   uint
+	QuotaBackendBytes int64
 }
 
 // TLSInfo wraps etcd transport TLSInfo
@@ -60,6 +73,8 @@ type TLSInfo transport.TLSInfo
 func NewConfig() *Config {
 	c := &Config{}
 	c.DataDir = path.SystemCacheDir("sensu-backend")
+	c.MaxRequestBytes = DefaultMaxRequestBytes
+	c.QuotaBackendBytes = DefaultQuotaBackendBytes
 
 	return c
 }
@@ -94,6 +109,11 @@ type Etcd struct {
 // BackendID returns the ID of the etcd cluster member
 func (e *Etcd) BackendID() (result string) {
 	return e.etcd.Server.ID().String()
+}
+
+// GetClusterVersion returns the cluster version of the etcd server
+func (e *Etcd) GetClusterVersion() string {
+	return e.etcd.Server.ClusterVersion().String()
 }
 
 // NewEtcd returns a new, configured, and running Etcd. The running Etcd will
@@ -149,6 +169,8 @@ func NewEtcd(config *Config) (*Etcd, error) {
 	cfg.LPUrls = lpURLs
 	cfg.PeerTLSInfo = (transport.TLSInfo)(config.PeerTLSInfo)
 
+	cfg.CipherSuites = config.CipherSuites
+
 	// Cluster config
 	cfg.InitialClusterToken = config.InitialClusterToken
 	cfg.InitialCluster = config.InitialCluster
@@ -158,8 +180,8 @@ func NewEtcd(config *Config) (*Etcd, error) {
 	// revision.
 	cfg.AutoCompactionMode = "revision"
 	cfg.AutoCompactionRetention = "2"
-	// Default to 4G etcd size. TODO: make this configurable.
-	cfg.QuotaBackendBytes = int64(4 * 1024 * 1024 * 1024)
+	cfg.QuotaBackendBytes = config.QuotaBackendBytes
+	cfg.MaxRequestBytes = config.MaxRequestBytes
 
 	capnslog.SetFormatter(NewLogrusFormatter())